pyproject.toml
[project]
name = "bajaj-rag"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "pymupdf (>=1.26.3,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "google-cloud-aiplatform (>=1.106.0,<2.0.0)",
    "google-auth (>=2.40.3,<3.0.0)",
    "pinecone (>=7.3.0,<8.0.0)",
    "langchain-google-vertexai (>=2.0.28,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
black = "^25.1.0"
isort = "^6.0.1"

